// src/shared/sdk/chronos/models/Service.ts
/**
 * Chronos API
 * API documentation for the Chronos application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Service {
    'categoryId'?: number | null;
    'color'?: string | null;
    'createdAt': Date;
    'currency': string;
    'description'?: string | null;
    'durationMinutes': number;
    'id': number;
    'isActive': boolean;
    'name': string;
    'priceCents': number;
    'tenantId': number;
    'updatedAt': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "categoryId",
            "baseName": "category_id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "color",
            "baseName": "color",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "durationMinutes",
            "baseName": "duration_minutes",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "isActive",
            "baseName": "is_active",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "priceCents",
            "baseName": "price_cents",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "tenantId",
            "baseName": "tenant_id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return Service.attributeTypeMap;
    }

    public constructor() {
    }
}